backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.6.0
dnspython>=2.4.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
pydantic>=2.5.0
gunicorn>=21.0.0
PyJWT>=2.8.0
werkzeug>=3.0.0
certifi>=2023.0.0
